requirements.txt
pyyaml
python-daemon